feat: `secondary check` asks one secondary what doctor asks all of them
`bindizr secondary check <name>` and `POST /secondaries/{name}/check`
resolve the registered address, compare the catalog zone serial the
secondary serves with the one Bindizr's own listener serves, and send it
a NOTIFY for the catalog zone; the CLI exits non-zero when any part fails.

The diagnostics now share one vocabulary: `SecondaryStatusResponse`
carries the classification (`in_sync`, `lagging`, `ahead`, `unreachable`,
and `reachable` when Bindizr's own serial is unknown) for `zone status`,
`doctor`, and the check alike, serials are u32 throughout, and
`NotifyCheckResponse` reports each NOTIFY sent. `loopback_if_unspecified`
moves to core so the service can probe the listener. CLAUDE.md gains the
status/check/doctor naming rule, the secondaries docs lead into the check
from the parts it reports on, and the OpenAPI document is regenerated.

+++ crates/bindizr/src/cli/output/table.rs
@@ -408,7 +408,7 @@ impl SecondaryStatusRow {
     }
 
     /// Build a table row from a secondary server's status.
-    fn from_secondary(secondary: &SecondaryStatusResponse, zone_serial: i32) -> Self {
+    fn from_secondary(secondary: &SecondaryStatusResponse, zone_serial: u32) -> Self {
         let detail = match secondary.error.as_deref() {
             Some(error) if secondary.is_unreachable() => {
                 format!("{} ({})", secondary.status, error)
@@ -423,7 +423,7 @@ impl SecondaryStatusRow {
                 .map_or_else(|| "-".to_string(), |serial| serial.to_string()),
             lag: secondary.visible_serial.map_or_else(
                 || "-".to_string(),
-                |serial| (i64::from(zone_serial) - serial).to_string(),
+                |serial| (i64::from(zone_serial) - i64::from(serial)).to_string(),
             ),
         }
     }
